requirements.txt
accelerate==0.23.0
coloredlogs==15.0.1
einops==0.7.0
ema_pytorch==0.2.2
imageio==2.31.3
matplotlib==3.7.1
numpy==1.24.2
open3d==0.17.0
opencv_python==4.7.0.72
Pillow
pytorch_fid==0.3.0
scipy==1.13.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.65.0
